+++ fabrictestbed_extensions/fablib/node.py
@@ -62,6 +62,7 @@ from fim.user import NodeType
 from IPython.core.display_functions import display
 from tabulate import tabulate
 
+from fabrictestbed_extensions.fablib.constants import Constants
 from fabrictestbed_extensions.fablib.network_service import NetworkService
 
 if TYPE_CHECKING:
@@ -752,8 +753,13 @@ class Node:
         :param username: Optional username parameter.  The username
             likely should be picked to match the image type.
         """
+        if self.get_fim_node().type == NodeType.Switch and not username:
+            self.username = Constants.FABRIC_USER
+            return
         if username is not None:
             self.username = username
+        elif "default_centos10_stream" == self.get_image():
+            self.username = "cloud-user"
         elif "default_centos9_stream" == self.get_image():
             self.username = "cloud-user"
         elif "centos" in self.get_image():
@@ -2352,7 +2358,7 @@ class Node:
                   result of ``ip addr list``.
         """
         try:
-            if self.ip_addr_list_json is not None and update == False:
+            if self.ip_addr_list_json is not None and not update:
                 return self.ip_addr_list_json
             else:
                 if output == "json":